animations finished unfixed bug: light attack can play while interacting with other objects

+++ War of the Gods/Assets/Scripts/HealthBar.cs
@@ -10,14 +10,14 @@ namespace JP
         public Slider slider;
 
         // Set Health Bar Slider to Max Value
-        public void SetMaxHealth(int maxHealth)
+        public void SetMaxHealth(float maxHealth)
         {
             slider.maxValue = maxHealth;
             slider.value = maxHealth;
         }
 
         // Set Health Bar to current Health value
-        public void SetCurrentHealth(int currentHealth)
+        public void SetCurrentHealth(float currentHealth)
         {
             slider.value = currentHealth;
         }
